% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_isd_sites.R
\name{create_smonitor_site_code}
\alias{create_smonitor_site_code}
\title{Function to convert an ISD site code to a \strong{smonitor} site code.}
\usage{
create_smonitor_site_code(site, df = NA)
}
\arguments{
\item{site}{Vector of sites to clean.}

\item{df}{Tibble from \code{get_isd_sites} if already retrieved.}
}
\value{
Tibble.
}
\description{
Function to convert an ISD site code to a \strong{smonitor} site code.
}
\author{
Stuart K. Grange
}
