% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dowloadISD.R
\name{downloadISD}
\alias{downloadISD}
\title{Download Integrated Surface Data from NOAA}
\usage{
downloadISD(usaf, wban, year, parameters)
}
\arguments{
\item{usaf}{The USAF master station catalog identifier.}

\item{wban}{The NCEI WBAN station identifier.}

\item{year}{The year of data.}

\item{parameters}{A vector of parameters to download. For a list of parameters, 
see the \code{SHORT_NAME} column in \code{data(ISD_documentation)}. If missing,
all parameters are downloaded.}
}
\description{
Download Integrated Surface Data from NOAA
}

