% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_isd_data.R
\name{get_isd_data}
\alias{get_isd_data}
\title{Function to directly get/import ISD data.}
\usage{
get_isd_data(
  site,
  year,
  priority = FALSE,
  longer = FALSE,
  parallel = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{site}{Vector of ISD site codes.}

\item{year}{Vector of years.}

\item{priority}{Should only a selected priority group of variables be 
returned?}

\item{longer}{Should the observations be reshaped to be in "long" format?}

\item{parallel}{Should files be read in parallel? If \code{TRUE}, a 
\strong{futures} back-end must be registered.}

\item{verbose}{Should the function give messages?}
}
\value{
Tibble.
}
\description{
Function to directly get/import ISD data.
}
\examples{

# Load an example data file
get_isd_data(site = "11290099999", year = 2020, priority = TRUE)

}
\seealso{
\href{https://www.ncdc.noaa.gov/isd}{ISD}
}
\author{
Stuart K. Grange.
}
