% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_isd_sites.R
\name{get_isd_sites}
\alias{get_isd_sites}
\title{Function to get metadata for the NOAA's Integrated Surface Database (ISD) 
sites.}
\usage{
get_isd_sites(clean = TRUE)
}
\arguments{
\item{clean}{Should the returned tibble be cleaned in the \strong{smonitor} 
style?}
}
\value{
Tibble.
}
\description{
Function to get metadata for the NOAA's Integrated Surface Database (ISD) 
sites.
}
\examples{

# Get isd sites data
get_isd_sites()

}
\seealso{
\href{https://www.ncdc.noaa.gov/isd}{ISD}
}
\author{
Stuart K. Grange
}
