% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getISDhistory.R
\name{getISDhistory}
\alias{getISDhistory}
\title{Download the latest version of the station history file}
\usage{
getISDhistory()
}
\description{
Downloads the file found at \url{ftp://ftp.ncdc.noaa.gov/pub/data/noaa/isd-history.csv}.
The table lists all stations, their location, and their period of activity.
}

