% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_isd_data.R
\name{isd_build_urls}
\alias{isd_build_urls}
\title{Function to build ISD remote file names.}
\usage{
isd_build_urls(site, year)
}
\arguments{
\item{site}{Vector of ISD site codes.}

\item{year}{Vector of years.}
}
\value{
Character vector.
}
\description{
Function to build ISD remote file names.
}
\examples{

isd_build_urls(c("08495099999", "03334099999", "08536099999"), 2000:2002)

}
\author{
Stuart K. Grange.
}
