% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isd_download_data.R
\name{isd_download_data}
\alias{isd_download_data}
\title{Function to download ISD data files.}
\usage{
isd_download_data(file_remote, file_local, parallel = FALSE, verbose = FALSE)
}
\arguments{
\item{file_remote}{Vector of remote file names.}

\item{file_local}{Vector of file names to export data to.}

\item{parallel}{Should files be read in parallel? If \code{TRUE}, a 
\strong{futures} back-end must be registered.}

\item{verbose}{Should the function give messages?}
}
\value{
Invisible NULL.
}
\description{
Function to download ISD data files.
}
\examples{

\dontrun{

# Download some files
isd_download_data(
  file = "https://www.ncei.noaa.gov/data/global-hourly/access/2020/11290099999.csv", 
  directory_output = "~/Desktop"
)

}

}
\seealso{
\href{https://www.ncdc.noaa.gov/isd}{ISD}
}
\author{
Stuart K. Grange.
}
