% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isd_file_name_local.R
\name{isd_file_name_local}
\alias{isd_file_name_local}
\title{Function to build local file name from ISD data file URL.}
\usage{
isd_file_name_local(file, directory = NA, extension = "rds")
}
\arguments{
\item{file}{Vector of remote ISD data file URLs.}

\item{directory}{Directory to add to file name (optional).}

\item{extension}{File type extension to append.}
}
\value{
\strong{fs} path.
}
\description{
Function to build local file name from ISD data file URL.
}
\examples{

isd_file_name_local(
  "https://www.ncei.noaa.gov/data/global-hourly/access/2018/11290099999.csv"
)

isd_file_name_local(
  "https://www.ncei.noaa.gov/data/global-hourly/access/2018/11290099999.csv",
  directory = "~/Desktop"
)

}
\author{
Stuart K. Grange
}
