% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ISD_lookup}
\alias{ISD_lookup}
\title{Lookup table for coded values in NOAA's Integrated Surface Data}
\format{An object of class \code{data.frame} with 186 rows and 3 columns.}
\source{
\url{ftp://ftp.ncdc.noaa.gov/pub/data/noaa/ish-format-document.pdf}
\describe{
  \item{FIELD}{The name of the column in the downloaded weather data.}
  \item{CODE}{The character value found in the coded weather data.}
  \item{DESCRIPTION}{The corresponding value.}
}
}
\usage{
ISD_lookup
}
\description{
The \code{ISDr} package facilitates the downloading of fixed length files
from the Integrated Surface Database via \url{ftp://ftp.ncdc.noaa.gov/pub/data/noaa}.
This data set contains the key for coded values in the downloaded data.
}
\keyword{datasets}

