% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isd_pad_site.R
\name{isd_pad_site}
\alias{isd_pad_site}
\title{Function to pad ISD site code.}
\usage{
isd_pad_site(x)
}
\arguments{
\item{x}{Vector of ISD sites.}
}
\value{
Character vector.
}
\description{
Function to pad ISD site code.
}
\examples{

isd_pad_site(03772099999)

}
\author{
Stuart K. Grange and Nelson the Duck.
}
