% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_isd.R
\name{isd_read}
\alias{isd_read}
\title{Function to read ISD data files.}
\usage{
isd_read(
  file,
  priority = FALSE,
  longer = FALSE,
  parallel = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{file}{Vector of file names.}

\item{priority}{Should only a selected priority group of variables be 
returned?}

\item{longer}{Should the observations be reshaped to be in "long" format?}

\item{parallel}{Should files be read in parallel? If \code{TRUE}, a 
\strong{futures} back-end must be registered.}

\item{verbose}{Should the function give messages?}
}
\value{
Tibble.
}
\description{
Function to read ISD data files.
}
\examples{

# Load an example data file
isd_read(
  "https://www.ncei.noaa.gov/data/global-hourly/access/2020/11290099999.csv"
)

isd_read(
  "https://www.ncei.noaa.gov/data/global-hourly/access/2020/11290099999.csv",
  priority = TRUE
)

}
\seealso{
\href{https://www.ncdc.noaa.gov/isd}{ISD}
}
\author{
Stuart K. Grange.
}
