% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_isd_sites.R
\name{isd_site_to_smonitor_site}
\alias{isd_site_to_smonitor_site}
\title{Function clean an ISD site code to a \strong{smonitor} site code.}
\usage{
isd_site_to_smonitor_site(site_isd, country)
}
\arguments{
\item{site_isd}{A vector of ISD site codes.}

\item{country}{Country suffix. This will generally be the ISO code.}
}
\value{
Character vector.
}
\description{
Function clean an ISD site code to a \strong{smonitor} site code.
}
\author{
Stuart K. Grange
}
