% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recodeISD.R
\name{recodeISD}
\alias{recodeISD}
\title{Recode Integrated Surface Data (ISD)}
\usage{
recodeISD(ISD_readings, columns)
}
\description{
Some columns of the data are coded. The codes can be found in
 \code{data(ISD_lookup)}. This function will replace the codes with the values
 found in the \code{DESCRIPTION} column of \code{ISD_lookup}.
 @param ISD_readings A \code{data.frame} of Integrated Surface Data.
 @param columns A character vector of columns to be recoded. If missing,
 all coded columns will be recoded.
}

