% Generated by roxygen2 (4.0.2): do not edit by hand
\name{downloadISD}
\alias{downloadISD}
\title{Download Integrated Surface Data}
\usage{
downloadISD(state, years)
}
\arguments{
\item{state}{two letter abbreviation of state to be downloaded}

\item{years}{a numeric vector of beginning and ending years (e.g. c(2005, 2010))}
}
\description{
Download integrated surface data from NOAA/NCDC into SQLite database
}
\examples{
downloadISD(state = "IN", years = c(2007, 2012))
}

